Fix dual masters race condition.
After a successful failover and the backup is active, a dual masters
exception could be thrown if:

1) a client is trying to connect to the primary server
2) the primary server is restarted before the client times out

The primary server would take the client connection as an event to
turn into a master, unaware that the backup was still active.

This patch adds a timeout - the primary server will not respond to
clients until two heartbeats have passed, allowing time for the backup
to notify the primary if it's active.