Update Go examples to build like the others.
Fixes #305.

The gozmq package API was updated in a way that was necessarily
backwards incompatible: Poll must receive a time.Duration instead of an
int64.  This commit updates the examples to use the new API.

Other example directories have build scripts and require no spelunking
through code and/or build errors to get things working.  This commit
also updates the Go examples to build so simply.