Correct invalid unicode character in "theyre", which is now "they're"
It took me a while to realize that this issue was not a problem of my toolchain, but an actual
issue in the document itself, which would probably prevent it from compiling to PDF on
other systems as well.

Now it is verified to build correctly to PDF and ePub on a ubuntu system.