Remove redundant static_cast<void*> in examples.
Not only is it unnecessary because zmq::socket_t automatically converts
to void* via operator void*, it is also misleading as the operator is
called in spite of the appearance that it's just a cast.

Now, other comments in zmq.hpp indicate that this is not really the
intent (note the comments near the bool conversion), and I believe the
cleanest fix would be to add a member to zmq::socket_t that returns the
void* needed for polling (.pollable()? I really can't find a good name),
or for zmq::pollitem_t to handle the conversion transparently.