Update hwserver.c for readability -- tie related constants together, expand abbreviations
There were several constants in the code that were magic numbers. The new code makes it clear which ones are dependent on each other. I also replaced "rc" with "response" to avoid a convention that not everyone may know. 

I do wish that the "0" for flags in zmq_recv and zmq_send had a proper #define somewhere. If there was a `#define ZMQ_NOFLAGS 0` then it would be easier for people to look up the header file containing all the flag options that are valid for these functions from code. But that doesn't exist, so I left the zeros.